+++ packages/cli/src/colors/index.ts
@@ -1,4 +1,4 @@
-export * from './colorMetadata.ts';
 export * from './scale.ts';
+export * from './specs.ts';
 export * from './types.ts';
 export * from './utils.ts';
