+++ packages/cli/src/colors/specs.ts
